Update `package.xml` and `CMakelists.txt`
to reflect the settings that make sense from those in the `Universal_Robots_ROS2_Description` package.
For `CMakelists.txt` I commented out the sections that where auto-generated by the `ros2 pkg create` command, but not needed at the moment. I commented them out so that we don't loose them to posterity.
For `package.xml` it is mainly additions.