add lines from `ros2_robotiq_driver` package to `package.xml` and `CMakelists.txt`
to match the latter's repo's files as much as possible. We keep details specific to package name and author as generated in last commit, but otherwise add things that later have to be renamed or doesn't exist yet.
Note that the addition to the package from the `ros2_robotiq_driver` package is indeed purely _additive_ with respect to the `ros2 pkg create` generated files (bar a three line comment), indicating they ran the same command and added extra lines to suit their needs.